fix vertical / horizontal

+++ src/exec/utils.c
@@ -6,7 +6,7 @@
 /*   By: rgallien <rgallien@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/03 20:42:34 by rgallien          #+#    #+#             */
-/*   Updated: 2024/10/25 17:34:28 by rgallien         ###   ########.fr       */
+/*   Updated: 2024/10/27 16:33:52 by rgallien         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 #define M_PI 3.14159265358979323846
 #endif
 
-double	found_distance(int x1, int y1, int x2, int y2)
+double	found_distance(double x1, double y1, double x2, double y2)
 {
 	if (x1 == x2 && y1 == y2)
 		return (39 * 50);
